Use `SettingsRepository` to retrieve the "schedule refresh interval" setting
