+++ app/src/main/java/nerd/tuxmobil/fahrplan/congress/preferences/RealSharedPreferencesRepository.kt
@@ -4,7 +4,6 @@ import android.annotation.SuppressLint
 import android.content.Context
 import android.content.SharedPreferences
 import androidx.core.content.edit
-import nerd.tuxmobil.fahrplan.congress.R
 
 class RealSharedPreferencesRepository(val context: Context) : SharedPreferencesRepository {
 
@@ -27,17 +26,6 @@ class RealSharedPreferencesRepository(val context: Context) : SharedPreferencesR
     }
 
     private val preferences: SharedPreferences = DefaultSettingsRepository.getDefaultSharedPreferences(context.applicationContext)
-
-    override fun getScheduleRefreshIntervalDefaultValue(): Int {
-        return context.getString(R.string.preference_default_value_schedule_refresh_interval_value).toInt()
-    }
-
-    override fun getScheduleRefreshInterval(): Int {
-        val key = context.getString(R.string.preference_key_schedule_refresh_interval_index)
-        val defaultValue = context.getString(R.string.preference_default_value_schedule_refresh_interval_value)
-        val value = preferences.getString(key, defaultValue)!!
-        return value.toInt()
-    }
 
     override fun getDisplayDayIndex() = preferences.getInt(DISPLAY_DAY_INDEX_KEY, 1)
